The model is rebuilt with each window normalised
The model is rebuilt with each window normalized and window size is kept to 200. The model can now detect each dip of packets but fails to predict the rise of packets as the window size doesn't clear it. Increasing window size may have served the purpose but lack of data is a constraint.